requirements.txt
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
Requests==2.32.3
tqdm==4.67.0
